+++ WorldBankAPIver2/src/main/java/dataFetchers/worldBankFetcher.java
@@ -15,8 +15,8 @@ public class worldBankFetcher implements Fetcher{
 	
 	public JsonArray fetchData(UserSelection selection, String analysisTypeCode) {
 		String country = selection.getCountryCode();
-		int fromYear = selection.getFromYear();
-		int toYear = selection.getToYear();
+		long fromYear = selection.getFromYear();
+		long toYear = selection.getToYear();
 		
 		String urlString = String.format(""
 				+ "\nhttp://api.worldbank.org/v2/country/%s/indicator/%s?date=%d:%d&format=json"
